fix build

+++ pkg/square/square_fuzz_test.go
@@ -42,7 +42,7 @@ func FuzzSquare(f *testing.F) {
 		require.NoError(t, err)
 		require.True(t, s.Equals(s2))
 
-		cacher := inclusion.NewSubtreeCacher(s.Size())
+		cacher := inclusion.NewSubtreeCacher(uint64(s.Size()))
 		eds, err := rsmt2d.ComputeExtendedDataSquare(shares.ToBytes(s), appconsts.DefaultCodec(), cacher.Constructor)
 		require.NoError(t, err)
 		dah := da.NewDataAvailabilityHeader(eds)
